# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wheels API",
        "description": "API for wheels app",
        "contact": {
            "email": "adrian.sullca@cirvianum.cat"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/profile/{id}/stats": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get profile statistics for a user",
                "description": "Returns the statistics of a user's profile, including announcement count, review count, and total favorites.",
                "operationId": "9a85fa6f69b0f5d47dae871f53629ea4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user whose profile statistics are being fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile statistics fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "John Doe",
                                                "email": "john@example.com"
                                            }
                                        },
                                        "stats": {
                                            "properties": {
                                                "announcements_count": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "reviews_count": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total_favorites": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required"
                    }
                }
            }
        },
        "/api/profile/{id}/announcements": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get all announcements by a user",
                "description": "Fetches a list of announcements made by a user. If the authenticated user is not the profile owner, only active announcements are returned.",
                "operationId": "4e2f2f191f0193d1b05a10b54f812a48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user whose announcements are being fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's announcements fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userAuthenticated": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "announcements": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Car for Sale"
                                                    },
                                                    "photoUrls": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "https://example.com/photo1.jpg"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required"
                    }
                }
            }
        },
        "/api/profile/{id}/reviews": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get reviews for a user",
                "description": "Fetches all reviews left for a user. Includes the details of the users who left the reviews.",
                "operationId": "10e362897ee3f580934908f5e86e62a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user whose reviews are being fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User reviews fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reviews": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "example": {
                                                    "id": 1,
                                                    "rating": 5,
                                                    "comment": "Great user!",
                                                    "valuator": {
                                                        "id": 2,
                                                        "name": "Jane Doe"
                                                    }
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required"
                    }
                }
            }
        },
...
    },
...
}